+++ protobuf/lib/meta.dart
@@ -52,6 +52,7 @@ const GeneratedMessage_reservedNames = <String>[
   'toBuilder',
   'toDebugString',
   'toProto3Json',
+  'toProto3JsonString',
   'toString',
   'unknownFields',
   'writeToBuffer',
@@ -79,6 +80,7 @@ const GeneratedMessage_reservedNames = <String>[
   r'$_setSignedInt32',
   r'$_setString',
   r'$_setUnsignedInt32',
+  r'$_toProto3JsonSink',
   r'$_whichOneof',
 ];