+++ protobuf/lib/protobuf.dart
@@ -8,12 +8,12 @@
 library protobuf;
 
 import 'dart:collection' show ListBase, MapBase;
-import 'dart:convert'
-    show base64Decode, base64Encode, jsonEncode, jsonDecode, Utf8Codec;
+import 'dart:convert' show base64Decode, base64Encode, jsonDecode, Utf8Codec;
 import 'dart:math' as math;
 import 'dart:typed_data' show TypedData, Uint8List, ByteData, Endian;
 
 import 'package:fixnum/fixnum.dart' show Int64;
+import 'package:jsontool/jsontool.dart';
 import 'package:meta/meta.dart' show UseResult;
 
 import 'src/protobuf/json_parsing_context.dart';
@@ -38,11 +38,13 @@ part 'src/protobuf/field_set.dart';
 part 'src/protobuf/field_type.dart';
 part 'src/protobuf/generated_message.dart';
 part 'src/protobuf/generated_service.dart';
-part 'src/protobuf/json.dart';
+part 'src/protobuf/json_reader.dart';
+part 'src/protobuf/json_writer.dart';
 part 'src/protobuf/pb_list.dart';
 part 'src/protobuf/pb_map.dart';
 part 'src/protobuf/protobuf_enum.dart';
-part 'src/protobuf/proto3_json.dart';
+part 'src/protobuf/proto3_json_reader.dart';
+part 'src/protobuf/proto3_json_writer.dart';
 part 'src/protobuf/rpc_client.dart';
 part 'src/protobuf/unknown_field_set.dart';
 part 'src/protobuf/utils.dart';